+++ hittable.h
@@ -11,7 +11,7 @@ struct hit_record
 {
 	point3 p;//交点
 	vec3 normal;//法线
-	double t;//距离
+	float t;//距离
 	bool front_face;//外面
 	shared_ptr<material> mat_ptr;//程序开始时候，光线碰撞到表面时，则会将此指针指向球类里的hit_record指针
 
@@ -27,7 +27,7 @@ class hittable
 {
 public:
 	//当t位于区间（t_min，t_max）内时，则记录下光线击中点的信息。 //用纯虚函数，方便后面不同的继承类更具自身属性更改
-	virtual bool hit(const ray& r, double t_min, double t_max, hit_record& record) const = 0;
+	virtual bool hit(const ray& r, float t_min, float t_max, hit_record& record) const = 0;
 };
 
 #endif // !HITTABLE_H
